+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Changes
+### Changes
 - Improve `PandasIO` performance (~4x faster for large datasets).
 - Typing updates (notably numpy `2.0`).
 - Added `concurrent_operation_action=raise|ignore` to `AbstractFether`. Default