+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2024-09-28
+
 ### Added
 - Property `TranslationMap.len_per_source`.
 
@@ -310,7 +312,8 @@ cookiecutter template.
 - Switch to relative imports.
 - Fix some intersphinx issues.
 
-[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.11.1...HEAD
+[Unreleased]: https://github.com/rsundqvist/id-translation/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/rsundqvist/id-translator/compare/v0.11.1...v0.12.0
 [0.11.1]: https://github.com/rsundqvist/id-translator/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/rsundqvist/id-translator/compare/v0.10.2...v0.11.0
 [0.10.2]: https://github.com/rsundqvist/id-translator/compare/v0.10.1...v0.10.2